Simplify session renewal
