+++ prisma/schema.prisma
@@ -6,9 +6,9 @@ generator client {
 }
 
 datasource db {
-  provider  = "postgresql"
-  url       = env("POSTGRES_PRISMA_URL")
-  directUrl = env("POSTGRES_URL_NON_POOLING")
+  provider          = "postgresql"
+  url               = env("POSTGRES_PRISMA_URL")
+  directUrl         = env("POSTGRES_URL_NON_POOLING")
   // shadowDatabaseUrl = env("POSTGRES_URL_NON_POOLING")
 }
 
@@ -31,7 +31,6 @@ model MemberToken {
   memberUuid  String
   invalidated Boolean  @default(false)
   createdAt   DateTime @default(now())
-  expiresAt   DateTime
 
   member Member @relation(fields: [memberUuid], references: [uuid], onDelete: Cascade)
 }
