--- src/lib/auth/session.ts
@@ -1,4 +1,0 @@
-export const SESSION_COOKIE_NAME = 'session.v2';
-export const TOKEN_TTL_SECONDS = 60 * 60 * 24 * 30;
-export const TOKEN_TTL_MS = TOKEN_TTL_SECONDS * 1000;
-export const TOKEN_REFRESH_THRESHOLD_MS = 1000 * 60 * 60 * 24 * 7;
