+++ src/lib/server/api/v2/auth/index.ts
@@ -1,4 +1,3 @@
 export { login, type LoginRequestData, type LoginResponseData } from './login';
 export { loginConfirm, type LoginConfirmRequestData, type LoginConfirmResponseData } from './loginConfirm';
 export { register, type RegisterRequestData, type RegisterResponseData } from './register';
-export { refreshToken, type RefreshTokenResponseData } from './refresh';
