+++ src/lib/server/api/v2/auth/login.ts
@@ -3,8 +3,6 @@ import { encryptRsa } from '$lib/data/crypto';
 import { db } from '$lib/server';
 import { checkStringParameter } from '$lib/utils';
 
-import { createMemberToken } from './token';
-
 export type LoginRequestData = {
   login: string;
 };
@@ -18,7 +16,17 @@ export async function login(data: LoginRequestData) {
     throw new ApiError(404, 'USER_NOT_FOUND', `User with login "${login}" not found`);
   }
 
-  const memberToken = await createMemberToken(member.uuid);
+  let memberToken = await db.memberToken.findFirst({
+    where: { AND: { memberUuid: member.uuid, invalidated: false } },
+  });
+  if (!memberToken) {
+    memberToken = await db.memberToken.create({
+      data: {
+        value: crypto.randomUUID(),
+        memberUuid: member.uuid,
+      },
+    });
+  }
 
   const publicKey = JSON.parse(member.publicKey) as JsonWebKey;
   const encryptedToken = await encryptRsa(publicKey, memberToken.value);
