+++ src/lib/server/api/v2/auth/loginConfirm.ts
@@ -4,7 +4,7 @@ import type { Cookies } from '@sveltejs/kit';
 
 import { checkStringParameter } from '$lib/utils';
 
-import { setSessionCookie } from './token';
+import { setSessionCookie } from './sessionCookie';
 
 export type LoginConfirmRequestData = {
   uuid: string;
@@ -15,21 +15,15 @@ export async function loginConfirm(data: LoginConfirmRequestData, cookies: Cooki
   const uuid = checkStringParameter(data.uuid, 'uuid');
   const token = checkStringParameter(data.token, 'token');
 
-  const memberToken = await db.memberToken.findFirst({
-    where: {
-      value: token,
-      invalidated: false,
-      expiresAt: { gt: new Date() },
-    },
-  });
+  const memberToken = await db.memberToken.findFirst({ where: { value: token, invalidated: false } });
 
   if (!memberToken || memberToken.memberUuid !== uuid) {
     throw new ApiError(401, 'UNAUTHORIZED', `Token "${token}" for user "${uuid}" not found`);
   }
 
-  setSessionCookie(cookies, token, memberToken.expiresAt);
+  setSessionCookie(cookies, token);
 
-  return { tokenExpiresAt: memberToken.expiresAt.toISOString() };
+  return true;
 }
 
 export type LoginConfirmResponseData = Awaited<ReturnType<typeof loginConfirm>>;
